+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added randomization of equipped Arcana Cards. Disabled by default, so this must be enabled in the config.
 
 ## [1.1.1] - 2024-06-03